+++ src/main/java/by/tamp/module3/hierarhy/SingleCar.java
@@ -2,8 +2,7 @@ package by.tamp.module3.hierarhy;
 
 public class SingleCar extends Car {
 
-    public SingleCar(String model, int price, int fuelConsumption, double weight) {
+    public SingleCar(String model, int price, int fuelConsumption, int weight) {
         super(model, price, fuelConsumption, weight);
-        //wings = new Wings("blue", 5);
     }
 }
